refactor: 코드 리뷰 반영 — 홈 사진 색인 VM 이동, 공용 카드 추출, 라이트 모드 고정
- SearchView의 홈 사진 색인 로딩을 SearchViewModel.loadPhotoIndex()로 이동
  (GetPhotosUseCase 주입, View는 photosById 상태만 소비)
- photosById 구성 시 uniquingKeysWith로 중복 photoId 방어
- ProfileCard/LogoutCard를 Resources/Components/GlassCards.swift로 공용화
- 다크모드 미지원 확정: label 색상 dark 블록 제거 + UIUserInterfaceStyle=Light
- SearchResultGrid 입력을 클로저 → dict 값으로 교체 (뷰 비교 가능성 개선)
