+++ Rephoto_iOS/Features/Search/Presentation/ViewModels/SearchViewModel.swift
@@ -10,16 +10,20 @@ import Foundation
 @Observable
 class SearchViewModel {
     let provider: SearchUseCaseProviderProtocol
-    
+    private let getPhotosUseCase: GetPhotosUseCaseProtocol
+
     var searchResults: [SearchResult] = []
+    /// 검색 결과(photoId)를 사진 상세용 전체 Photo로 매핑하기 위한 홈 사진 색인
+    private(set) var photosById: [Int: Photo] = [:]
     private(set) var isLoading = false
     private(set) var errorMessage: String?
     var query: String = ""
-    
-    init(provider: SearchUseCaseProviderProtocol) {
+
+    init(provider: SearchUseCaseProviderProtocol, getPhotosUseCase: GetPhotosUseCaseProtocol) {
         self.provider = provider
+        self.getPhotosUseCase = getPhotosUseCase
     }
-    
+
     @MainActor
     func search(query: String) async {
         isLoading = true
@@ -31,5 +35,11 @@ class SearchViewModel {
             errorMessage = error.localizedDescription
         }
         isLoading = false
+    }
+
+    @MainActor
+    func loadPhotoIndex() async {
+        guard let photos = try? await getPhotosUseCase.execute() else { return }
+        photosById = Dictionary(photos.map { ($0.photoId, $0) }, uniquingKeysWith: { first, _ in first })
     }
 }
