+++ Rephoto_iOS/Features/Settings/Presentation/Views/SettingsView.swift
@@ -18,7 +18,7 @@ struct SettingsView: View {
     var body: some View {
         ScrollView {
             VStack(alignment: .leading, spacing: 0) {
-                SettingsProfileCard(name: session.name)
+                ProfileCard(name: session.name, showsChevron: true)
 
                 Text("정보")
                     .font(.system(size: 13, weight: .medium))
@@ -45,40 +45,6 @@ struct SettingsView: View {
                 Task { await session.logout() }
             }
         }
-    }
-}
-
-// MARK: - SettingsProfileCard
-
-/// 아바타(이름 첫 글자) + 이름 글래스 프로필 카드
-private struct SettingsProfileCard: View {
-    let name: String
-
-    var body: some View {
-        HStack(spacing: 14) {
-            Text(name.prefix(1))
-                .font(.system(size: 22, weight: .bold))
-                .foregroundStyle(.white)
-                .frame(width: 52, height: 52)
-                .background(
-                    LinearGradient(colors: [.lightGreen, .mainGreen], startPoint: .top, endPoint: .bottom),
-                    in: Circle()
-                )
-
-            Text(name)
-                .font(.system(size: 18, weight: .bold))
-                .foregroundStyle(.labelPrimary)
-
-            Spacer()
-
-            Image(systemName: "chevron.right")
-                .font(.system(size: 14, weight: .semibold))
-                .foregroundStyle(.labelTertiary)
-        }
-        .padding(.leading, 14)
-        .padding(.trailing, 16)
-        .frame(height: 80)
-        .glassEffect(.regular, in: RoundedRectangle(cornerRadius: 18))
     }
 }
 
@@ -117,25 +83,6 @@ private struct VersionRow: View {
         .padding(.trailing, 16)
         .frame(height: 54)
         .glassEffect(.regular, in: RoundedRectangle(cornerRadius: 16))
-    }
-}
-
-// MARK: - LogoutCard
-
-/// 로그아웃 글래스 카드 버튼
-private struct LogoutCard: View {
-    let action: () -> Void
-
-    var body: some View {
-        Button(action: action) {
-            Text("로그아웃")
-                .font(.system(size: 17, weight: .bold))
-                .foregroundStyle(Color(red: 0xE2 / 255, green: 0x33 / 255, blue: 0x2F / 255))
-                .frame(maxWidth: .infinity)
-                .frame(height: 54)
-                .glassEffect(.regular.interactive(), in: RoundedRectangle(cornerRadius: 16))
-        }
-        .buttonStyle(.plain)
     }
 }
