+++ Rephoto_iOS/Resources/Colors/Colors.xcassets/labelPrimary.colorset/Contents.json
@@ -11,24 +11,6 @@
         }
       },
       "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "alpha" : "1.000",
-          "blue" : "0x18",
-          "green" : "0x1B",
-          "red" : "0x1C"
-        }
-      },
-      "idiom" : "universal"
     }
   ],
   "info" : {
