+++ Rephoto_iOS/Resources/Colors/Colors.xcassets/labelSecondary.colorset/Contents.json
@@ -11,24 +11,6 @@
         }
       },
       "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "alpha" : "1.000",
-          "blue" : "0x61",
-          "green" : "0x6A",
-          "red" : "0x6E"
-        }
-      },
-      "idiom" : "universal"
     }
   ],
   "info" : {
