+++ Rephoto_iOS/Resources/Colors/Colors.xcassets/labelTertiary.colorset/Contents.json
@@ -11,24 +11,6 @@
         }
       },
       "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "alpha" : "1.000",
-          "blue" : "0x93",
-          "green" : "0x9D",
-          "red" : "0xA2"
-        }
-      },
-      "idiom" : "universal"
     }
   ],
   "info" : {
